+++ source/packages/@aws-accelerator/constructs/lib/aws-route-53-resolver/get-endpoint-addresses/index.ts
@@ -11,11 +11,9 @@
  *  and limitations under the License.
  */
 
-import * as AWS from 'aws-sdk';
-
-import { throttlingBackOff } from '@aws-accelerator/utils/lib/throttle';
+import { paginateListResolverEndpointIpAddresses, Route53ResolverClient } from '@aws-sdk/client-route53resolver';
 import { CloudFormationCustomResourceEvent } from '@aws-accelerator/utils/lib/common-types';
-AWS.config.logger = console;
+import { setRetryStrategy } from '@aws-accelerator/utils/lib/common-functions';
 
 /**
  * Get Route 53 resolver endpoint details - Lambda handler
@@ -40,29 +38,27 @@ export async function handler(event: CloudFormationCustomResourceEvent): Promise
   const region: string = event.ResourceProperties['region'];
   const endpointId: string = event.ResourceProperties['endpointId'];
   const solutionId = process.env['SOLUTION_ID'];
-  const resolverClient = new AWS.Route53Resolver({ region: region, customUserAgent: solutionId });
+  const resolverClient = new Route53ResolverClient({
+    customUserAgent: solutionId,
+    region,
+    retryStrategy: setRetryStrategy(),
+  });
 
   switch (event.RequestType) {
     case 'Create':
     case 'Update':
       const ipArray: { Ip: string }[] = [];
-      let nextToken: string | undefined = undefined;
-      do {
-        const page = await throttlingBackOff(() =>
-          resolverClient
-            .listResolverEndpointIpAddresses({ ResolverEndpointId: endpointId, NextToken: nextToken })
-            .promise(),
-        );
-
-        // Loop through IP addresses and push to array
+      for await (const page of paginateListResolverEndpointIpAddresses(
+        { client: resolverClient },
+        { ResolverEndpointId: endpointId },
+        region,
+      )) {
         for (const item of page.IpAddresses ?? []) {
           if (item.Ip) {
             ipArray.push({ Ip: item.Ip });
           }
         }
-
-        nextToken = page.NextToken;
-      } while (nextToken);
+      }
 
       return {
         PhysicalResourceId: endpointId,