+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore(runtime): updated node runtime to node 20
 - chore(securityhub): update sdk v3
 - chore(identity-center): update sdk v3
+- chore(route53resolver): update sdk v3
 
 ### Configuration Changes
 - chore(config): add s3 object level and additional network controls